Feat: todoModal 기간 컴포넌트 분리 및 타입 설정
