+++ src/types/TodoDateType.ts
@@ -1,0 +1,1 @@
+export type TodoDateType = 'start' | 'end';
